requirements.txt
# Core Django and API framework
Django==5.2.3
djangorestframework==3.16.0
django-filter==24.3
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Configuration and environment
python-decouple==3.8

# Database
psycopg2-binary==2.9.10
dj-database-url==3.0.1
sqlparse==0.5.3

# Excel and data processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Static files and deployment
whitenoise==6.7.0
gunicorn==21.2.0

# Timezone and localization
tzdata==2025.2

# Security and HTTP
requests>=2.31.0

# SMS and Communication
twilio>=8.0.0

# Development tools
setuptools